+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add 'key_operators' column to schema output.
+
 ## [0.16.0] - 2025-09-16
 
 ### Added